Refactor Company Profiles and bug fixes
