--- backend/vehicles/src/company-profiles/company-profile.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { CompanyProfileService } from './service/company-profile.service';
-import { CompanyProfileController } from './controller/company-profile.controller';
-import { Company } from './entities/company.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Company])],
-  controllers: [CompanyProfileController],
-  providers: [CompanyProfileService],
-})
-export class CompanyProfileModule {}
