+++ backend/vehicles/src/modules/company/company-profiles/company-profile.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Post, Body, Param, Put } from '@nestjs/common';
-import { CreateCompanyProfileDto } from '../dto/create-company-profile.dto';
-import { UpdateCompanyProfileDto } from '../dto/update-company-profile.dto';
-import { CompanyProfileService } from '../service/company-profile.service';
+import { CreateCompanyProfileDto } from './dto/create-company-profile.dto';
+import { UpdateCompanyProfileDto } from './dto/update-company-profile.dto';
+import { CompanyProfileService } from './company-profile.service';
 import {
   ApiCreatedResponse,
   ApiInternalServerErrorResponse,
@@ -10,9 +10,9 @@ import {
   ApiOkResponse,
   ApiTags,
 } from '@nestjs/swagger';
-import { DataNotFoundException } from '../../common/exception/data-not-found.exception';
-import { ExceptionDto } from '../../common/dto/exception.dto';
-import { ReadCompanyProfileDto } from '../dto/read-company-profile.dto';
+import { DataNotFoundException } from '../../../common/exception/data-not-found.exception';
+import { ExceptionDto } from '../../../common/dto/exception.dto';
+import { ReadCompanyProfileDto } from './dto/read-company-profile.dto';
 
 @ApiTags('Company Profile')
 @ApiNotFoundResponse({
