+++ backend/vehicles/src/modules/company/company-profiles/company-profile.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { CompanyProfileService } from './company-profile.service';
+import { CompanyProfileController } from './company-profile.controller';
+import { Company } from './entities/company.entity';
+import { Contact } from './entities/contact.entity';
+import { Address } from './entities/address.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Company, Address, Contact])],
+  controllers: [CompanyProfileController],
+  providers: [CompanyProfileService],
+})
+export class CompanyProfileModule {}
