+++ backend/vehicles/src/modules/company/company-profiles/entities/contact.entity.ts
@@ -5,8 +5,8 @@ import {
   JoinColumn,
   PrimaryGeneratedColumn,
 } from 'typeorm';
-import { Province } from '../../common/entities/province.entity';
-import { Base } from '../../common/entities/base.entity';
+import { Province } from '../../../../common/entities/province.entity';
+import { Base } from '../../../../common/entities/base.entity';
 import { AutoMap } from '@automapper/classes';
 import { Company } from './company.entity';
